requirements.txt
fastapi
uvicorn
gino
asyncpg
psycopg2


pydantic[email]
httpx

python-dotenv
pytest